+++ README.md
@@ -1,15 +1,25 @@
-# budget-money
-A Rocket Money transaction analyzer tool
+# Budget Money
 
-Rocket Money allows users to export their financial transactions to a CSV file. This tool takes in one or more of these CSVs in order to produce statistics and visualizations to help a user better understand their spending and achieve their budgeting goals.
+Budget Money (bmoney) is a budgeting tool that builds on top of Rocket Money transaction exports.
+
+---
+
+Rocket Money allows users to export their financial transactions to a CSV file. Rocket Money (through their partnered service Plaid) provide transactions up to two years ago.
+
+Budget Money (this tool) provides an easy way to  takes in one or more of these CSVs in order to produce statistics and visualizations to help a user better understand their spending and achieve their budgeting goals.
+
+- Merge multiple Rocket Money CSV export files into one highly portable JSONL file
+- Display and easily edit your data
+- Extends Rocket Money metadata:
+    - Custom categories
+    - "Shared expenses" to make it easier to separate expenses with partners
+- Metrics and visualization dashboard to see category spending habit in more detail
+- Export category spend data to Google Sheets 
 
 # Installation
 
-`pip install budget-money`
-
-# What can this tool do
+`pip install bmoney`
 
 
 
 
-1)
