+++ bmoney/constants.py
