+++ pyproject.toml
@@ -1,7 +1,10 @@
 [project]
-name = "budgetmoney"
-version = "0.1.0"
-description = "A Rocket Money transaction analyzer tool"
+name = "bmoney"
+version = "0.2.0"
+authors = [{name="David Skarbrevik"}]
+license = "MIT"
+license-files = ["LICENSE"]
+description = "Budget Money - powerful budgeting tools extending Rocket Money transaction exports"
 readme = "README.md"
 requires-python = ">=3.10, <3.14"
 dependencies = [
@@ -19,6 +22,10 @@ dependencies = [
     "typer>=0.15.1",
 ]
 
+[project.urls]
+Homepage = "https://github.com/dskarbrevik/budget-money"
+Issues = "https://github.com/dskarbrevik/budget-money/issues"
+
 [dependency-groups]
 dev = [
     "pytest>=8.3.4",
@@ -29,11 +36,11 @@ dev = [
 ]
 
 [project.scripts]
-bmoney = "budgetmoney.cli:app"
+bmoney = "bmoney.cli:app"
 
 [build-system]
 requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["budgetmoney"]
+packages = ["bmoney"]
